Start of graphQL backend
So this is kind of like a default setup I pulled from the nexus site but it gives me a good starting point to start messing around with stuff. We have basic functions of showing drafts/posts and also creating drafts/posts. Right now this stuff is stored in-memory so nothing will persist so next step is picking what database I actually want to use and implement that. Then after getting authentication setup and then hook up the flutter front end which will be interesting, I wanted to use apollo client but there is no official support for it so we will use graphql_flutter (or whatever the package is called). This will be a wild ride for me as everything here is extremely new to me so its a lot to take in.